Ensure rust code passes code standards
Adds a CI suite for building and linting rust code in the
dataplane. This minimizes and/or altogher blocks
non-idiomatic rust code from sneaking into the code base.

NOTE: Builds happen on a directory by directory basis and
are explicitly written as steps. Once we migrate the controller
from Go to Rust, we'll need to add a step for building that
as well.

+++ dataplane/ebpf/src/ingress/udp.rs
@@ -126,7 +126,7 @@ pub fn handle_udp_ingress(ctx: TcContext) -> Result<i32, i64> {
         next = 0;
     }
     unsafe {
-        GATEWAY_INDEXES.insert(&backend_key, &next, 0 as u64)?;
+        GATEWAY_INDEXES.insert(&backend_key, &next, 0_u64)?;
     }
 
     info!(&ctx, "redirect action: {}", action);
